Split datastore record tests into their own pytest-native file
DatasetRefURIsTestCase, StoredFileInfoTestCase and
TestDatastoreRecordTable are pure unit tests over the record types, so
they need nothing from the fixture plugin.

Test count conserved at 7 for the moved tests, 194 overall.
